+++ api/src/test/java/edu/utexas/tacc/tapis/files/integration/transfers/configs/TestFileTransfersConfig.java
@@ -1,0 +1,12 @@
+package edu.utexas.tacc.tapis.files.integration.transfers.configs;
+
+import java.util.List;
+
+public class TestFileTransfersConfig extends BaseTransfersIntegrationConfig {
+
+    List<TransfersConfig> transfers;
+
+    public List<TransfersConfig> getTransfers() {
+        return transfers;
+    }
+}